refactor(v3.2.4): 抽 651 行 panel_cards + special_cards
- panel_cards.py (183 行): GROUP_LABELS + render_jury_seat/chat/vote_bars/top3/risks
- special_cards.py (544 行): friendly_layer/fund_managers/panel_insights/school_scores/debate

## v3.2 累计 (assemble_report.py)

| commit | 抽离 | 行数变化 |
|---|---|---|
| v3.2.1 | svg_primitives.py (602 行) | 2964 → 2400 |
| v3.2.2 | dim_viz.py (742 行) | 2400 → 1708 |
| v3.2.3 | institutional.py (532 行) | 1708 → 1229 |
| v3.2.4 | panel/special (727 行) | 1229 → 587 |
| **total** | **2603 行 → 5 个子模块** | **-80%** |

## 向后兼容

assemble_report.py 对所有抽离函数做 re-export · 所有历史调用不变.
332 tests 全过 · 4 个 grep 式 test 扩展读 assemble + 对应 subset.
真机 e2e · 002217 assemble 0.0s 出报告 · 格式 100% 一致.
